+++ core/src/main/java/org/testcontainers/containers/DockerComposeContainer.java
@@ -7,6 +7,7 @@ import com.google.common.base.Joiner;
 import com.google.common.base.Splitter;
 import com.google.common.collect.Maps;
 import com.google.common.util.concurrent.Uninterruptibles;
+import lombok.NonNull;
 import org.apache.commons.lang.StringUtils;
 import org.apache.commons.lang.SystemUtils;
 import org.junit.runner.Description;
@@ -14,9 +15,11 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.slf4j.profiler.Profiler;
 import org.testcontainers.DockerClientFactory;
-import org.testcontainers.containers.output.OutputFrame;
 import org.testcontainers.containers.output.Slf4jLogConsumer;
 import org.testcontainers.containers.startupcheck.IndefiniteWaitOneShotStartupCheckStrategy;
+import org.testcontainers.containers.wait.strategy.Wait;
+import org.testcontainers.containers.wait.strategy.WaitAllStrategy;
+import org.testcontainers.containers.wait.strategy.WaitStrategy;
 import org.testcontainers.utility.*;
 import org.zeroturnaround.exec.InvalidExitValueException;
 import org.zeroturnaround.exec.ProcessExecutor;
@@ -26,8 +29,14 @@ import java.io.File;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.*;
+import java.time.Duration;
 import java.util.AbstractMap.SimpleEntry;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -60,6 +69,8 @@ public class DockerComposeContainer<SELF extends DockerComposeContainer<SELF>> e
 
     private final AtomicInteger nextAmbassadorPort = new AtomicInteger(2000);
     private final Map<String, Map<Integer, Integer>> ambassadorPortMappings = new ConcurrentHashMap<>();
+    private final Map<String, ComposeServiceWaitStrategyTarget> serviceInstanceMap = new ConcurrentHashMap<>();
+    private final Map<String, WaitAllStrategy> waitStrategyMap = new ConcurrentHashMap<>();
     private final SocatContainer ambassadorContainer = new SocatContainer();
 
     private static final Object MUTEX = new Object();
@@ -112,10 +123,8 @@ public class DockerComposeContainer<SELF extends DockerComposeContainer<SELF>> e
             }
             applyScaling(); // scale before up, so that all scaled instances are available first for linking
             createServices();
-            if (tailChildContainers) {
-                tailChildContainerLogs();
-            }
             startAmbassadorContainers(profiler);
+            waitUntilServiceStarted();
         }
     }
 
@@ -129,14 +138,34 @@ public class DockerComposeContainer<SELF extends DockerComposeContainer<SELF>> e
         runWithCompose("up -d");
     }
 
-    private void tailChildContainerLogs() {
-        listChildContainers().forEach(container ->
-                LogUtils.followOutput(dockerClient,
-                        container.getId(),
-                        new Slf4jLogConsumer(logger()).withPrefix(container.getNames()[0]),
-                        OutputFrame.OutputType.STDOUT,
-                        OutputFrame.OutputType.STDERR)
-        );
+    private void waitUntilServiceStarted() {
+        listChildContainers().forEach(this::createServiceInstance);
+        serviceInstanceMap.forEach(this::waitUntilServiceStarted);
+    }
+
+    private void createServiceInstance(Container container) {
+        String serviceName = getServiceNameFromContainer(container);
+        final ComposeServiceWaitStrategyTarget containerInstance = new ComposeServiceWaitStrategyTarget(container,
+            ambassadorContainer, ambassadorPortMappings.getOrDefault(serviceName, new HashMap<>()));
+
+        if (tailChildContainers) {
+            LogUtils.followOutput(DockerClientFactory.instance().client(), containerInstance.getContainerId(),
+                new Slf4jLogConsumer(logger()).withPrefix(container.getNames()[0]));
+        }
+       serviceInstanceMap.putIfAbsent(serviceName, containerInstance);
+    }
+
+    private void waitUntilServiceStarted(String serviceName, ComposeServiceWaitStrategyTarget serviceInstance) {
+        final WaitAllStrategy waitAllStrategy = waitStrategyMap.get(serviceName);
+        if(waitAllStrategy != null) {
+            waitAllStrategy.waitUntilReady(serviceInstance);
+        }
+    }
+
+    private String getServiceNameFromContainer(Container container) {
+        final String containerName = container.getLabels().get("com.docker.compose.service");
+        final String containerNumber = container.getLabels().get("com.docker.compose.container-number");
+        return String.format("%s_%s", containerName, containerNumber);
     }
 
     private void runWithCompose(String cmd) {
@@ -227,10 +256,20 @@ public class DockerComposeContainer<SELF extends DockerComposeContainer<SELF>> e
     }
 
     public SELF withExposedService(String serviceName, int servicePort) {
+        return withExposedService(serviceName, servicePort, Wait.defaultWaitStrategy());
+    }
 
-        if (!serviceName.matches(".*_[0-9]+")) {
-            serviceName += "_1"; // implicit first instance of this service
-        }
+    public DockerComposeContainer withExposedService(String serviceName, int instance, int servicePort) {
+        return withExposedService(serviceName + "_" + instance, servicePort);
+    }
+
+    public DockerComposeContainer withExposedService(String serviceName, int instance, int servicePort, WaitStrategy waitStrategy) {
+        return withExposedService(serviceName + "_" + instance, servicePort, waitStrategy);
+    }
+
+    public SELF withExposedService(String serviceName, int servicePort, @NonNull WaitStrategy waitStrategy) {
+
+        String serviceInstanceName = getServiceInstanceName(serviceName);
 
         /*
          * For every service/port pair that needs to be exposed, we register a target on an 'ambassador container'.
@@ -248,14 +287,44 @@ public class DockerComposeContainer<SELF extends DockerComposeContainer<SELF>> e
 
         // Ambassador container will be started together after docker compose has started
         int ambassadorPort = nextAmbassadorPort.getAndIncrement();
-        ambassadorPortMappings.computeIfAbsent(serviceName, __ -> new ConcurrentHashMap<>()).put(servicePort, ambassadorPort);
-        ambassadorContainer.withTarget(ambassadorPort, serviceName, servicePort);
-        ambassadorContainer.addLink(new FutureContainer(this.project + "_" + serviceName), serviceName);
+        ambassadorPortMappings.computeIfAbsent(serviceInstanceName, __ -> new ConcurrentHashMap<>()).put(servicePort, ambassadorPort);
+        ambassadorContainer.withTarget(ambassadorPort, serviceInstanceName, servicePort);
+        ambassadorContainer.addLink(new FutureContainer(this.project + "_" + serviceInstanceName), serviceInstanceName);
+        addWaitStrategy(serviceInstanceName, waitStrategy);
         return self();
     }
 
-    public DockerComposeContainer withExposedService(String serviceName, int instance, int servicePort) {
-        return withExposedService(serviceName + "_" + instance, servicePort);
+    private String getServiceInstanceName(String serviceName) {
+        String serviceInstanceName = serviceName;
+        if (!serviceInstanceName.matches(".*_[0-9]+")) {
+            serviceInstanceName += "_1"; // implicit first instance of this service
+        }
+        return serviceInstanceName;
+    }
+
+    /*
+     * can have multiple wait strategies for a single container, e.g. if waiting on several ports
+     * if no wait strategy is defined, the WaitAllStrategy will return immediately.
+     * The WaitAllStrategy uses an long timeout, because timeouts should be handled by the inner strategies.
+     */
+    private void addWaitStrategy(String serviceInstanceName, @NonNull WaitStrategy waitStrategy) {
+        final WaitAllStrategy waitAllStrategy = waitStrategyMap.computeIfAbsent(serviceInstanceName, __ ->
+            (WaitAllStrategy) new WaitAllStrategy().withStartupTimeout(Duration.ofMinutes(30)));
+        waitAllStrategy.withStrategy(waitStrategy);
+    }
+
+    /**
+     Specify the {@link WaitStrategy} to use to determine if the container is ready.
+     *
+     * @see org.testcontainers.containers.wait.strategy.Wait#defaultWaitStrategy()
+     * @param serviceName the name of the service to wait for
+     * @param waitStrategy the WaitStrategy to use
+     * @return this
+     */
+    public SELF waitingFor(String serviceName, @NonNull WaitStrategy waitStrategy) {
+        String serviceInstanceName = getServiceInstanceName(serviceName);
+        addWaitStrategy(serviceInstanceName, waitStrategy);
+        return self();
     }
 
     /**
@@ -283,7 +352,7 @@ public class DockerComposeContainer<SELF extends DockerComposeContainer<SELF>> e
      * @return a port that can be used for accessing the service container.
      */
     public Integer getServicePort(String serviceName, Integer servicePort) {
-        return ambassadorContainer.getMappedPort(ambassadorPortMappings.get(serviceName).get(servicePort));
+        return ambassadorContainer.getMappedPort(ambassadorPortMappings.get(getServiceInstanceName(serviceName)).get(servicePort));
     }
 
     public SELF withScaledService(String serviceBaseName, int numInstances) {