+++ core/src/main/java/org/testcontainers/containers/wait/strategy/WaitStrategy.java
@@ -1,0 +1,10 @@
+package org.testcontainers.containers.wait.strategy;
+
+import java.time.Duration;
+
+public interface WaitStrategy {
+
+    void waitUntilReady(WaitStrategyTarget waitStrategyTarget);
+
+    WaitStrategy withStartupTimeout(Duration startupTimeout);
+}