+++ modules/jdbc-test/src/test/java/org/testcontainers/junit/SimpleOracleTest.java
@@ -1,0 +1,39 @@
+package org.testcontainers.junit;
+
+import com.zaxxer.hikari.HikariConfig;
+import com.zaxxer.hikari.HikariDataSource;
+import org.junit.Rule;
+import org.junit.Test;
+import org.testcontainers.containers.OracleContainer;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
+
+import static org.rnorth.visibleassertions.VisibleAssertions.assertEquals;
+
+/**
+ * @author gusohal
+ */
+public class SimpleOracleTest {
+
+    @Rule
+    public OracleContainer oracle = new OracleContainer();
+
+    @Test
+    public void testSimple() throws SQLException {
+        HikariConfig hikariConfig = new HikariConfig();
+        hikariConfig.setJdbcUrl(oracle.getJdbcUrl());
+        hikariConfig.setUsername(oracle.getUsername());
+        hikariConfig.setPassword(oracle.getPassword());
+
+        HikariDataSource ds = new HikariDataSource(hikariConfig);
+        Statement statement = ds.getConnection().createStatement();
+        statement.execute("SELECT 1 FROM dual");
+        ResultSet resultSet = statement.getResultSet();
+
+        resultSet.next();
+        int resultSetInt = resultSet.getInt(1);
+        assertEquals("A basic SELECT query succeeds", 1, resultSetInt);
+    }
+}